Final clearout of `Node` as a public-facing object in pygambit (#1122)
This completes the removal of `Node` as a public-facing object.  Querying of extensive games is now done through the `History` object, the tuple of actions which constitute the course of play.
