+++ src/pygambit/gambit.pyx
@@ -172,11 +172,16 @@ class StrategyIndexedVector(_LabeledVector):
 
 @cython.cclass
 class NodeIndexedVector(_LabeledVector):
-    """A read-only mapping from a ``Node`` to a computed value, one entry per node.
+    """A read-only mapping from a node's History to a computed value, one entry
+    per node.
 
     Unlike ``PlayerIndexedVector``/``StrategyIndexedVector``, which are keyed by a stable
-    label, this is keyed by node identity: the value can genuinely differ between two
-    nodes, even nodes belonging to the same information set.
+    label, this is keyed by the node's own History -- a plain tuple of action labels
+    from the root, unique to that node -- since the value can genuinely differ between
+    two nodes, even nodes belonging to the same information set.
+
+    .. versionchanged:: 17.0.0
+        Keyed by a node's History rather than a ``Node`` object.
     """
     _label_kind = "node"
