+++ src/pygambit/qre.py
@@ -253,31 +253,23 @@ def _estimate_strategy_empirical(
     )
 
 
-def _selector_for_node(node: libgbt.Node) -> libgbt.H:
-    """A root-anchored `Selector` matching exactly `node`'s own path of action labels."""
-    labels = []
-    current = node
-    while current.parent is not None:
-        labels.append(current.prior_action.label)
-        current = current.parent
-    labels.reverse()
-    return libgbt.H.path(*labels)
-
-
 def _estimate_behavior_empirical(
         data: libgbt.MixedBehaviorProfile,
 ) -> LogitQREMixedBehaviorFitResult:
     flattened_data = [
-        data[_selector_for_node(node)][a]
+        data[libgbt.H.path(*history)][a]
         for p in data.game.players
-        for node in data.game.get_infosets(p)
-        for a in node.actions
+        for history in data.game.get_infosets(p)
+        for a in data.game.get_actions(libgbt.H.path(*history))
     ]
     normalized = data.normalize()
     regrets = [
-        [-normalized.action_regrets[node][a] for a in node.actions]
+        [
+            -normalized.action_regrets[libgbt.H.path(*history)][a]
+            for a in data.game.get_actions(libgbt.H.path(*history))
+        ]
         for player in data.game.players
-        for node in data.game.get_infosets(player)
+        for history in data.game.get_infosets(player)
     ]
     res = scipy.optimize.minimize(
         lambda x: -_empirical_log_like(x[0], regrets, flattened_data),
@@ -287,9 +279,10 @@ def _estimate_behavior_empirical(
     profile = data.game.mixed_behavior_profile()
     log_probs = iter(_empirical_log_logit_probs(res.x[0], regrets))
     for player in data.game.players:
-        for node in data.game.get_infosets(player):
-            profile[_selector_for_node(node)] = {
-                a: math.exp(next(log_probs)) for a in node.actions
+        for history in data.game.get_infosets(player):
+            selector = libgbt.H.path(*history)
+            profile[selector] = {
+                a: math.exp(next(log_probs)) for a in data.game.get_actions(selector)
             }
     return LogitQREMixedBehaviorFitResult(
         data, "empirical", res.x[0], profile, -res.fun
