+++ tests/test_behav.py
@@ -18,11 +18,27 @@ def _set_action_probs(profile: gbt.MixedBehaviorProfile, probs: list, rational_f
     """
     convert = (lambda p: gbt.Rational(p)) if rational_flag else (lambda p: p)
     probs_iter = iter(probs)
-    for infoset in games.all_infosets(profile.game):
-        node = next(iter(infoset.members))
-        profile[games.selector_for_node(node)] = {
-            a: convert(next(probs_iter)) for a in infoset.actions
+    game = profile.game
+    for infoset in games.all_infosets(game):
+        selector = gbt.H.path(*infoset)
+        profile[selector] = {
+            a: convert(next(probs_iter)) for a in game.get_actions(selector)
         }
+
+
+def _infoset_history(game: gbt.Game, label: str) -> tuple:
+    """The History of the representative member of the infoset historically
+    identified by `label`, matching the removed `Infoset`'s by-label lookup."""
+    return games._INFOSET_LABEL_HISTORIES[(game.title, label)]
+
+
+def _member_selector(game: gbt.Game, label: str) -> gbt.Selector:
+    """A `Selector` for some member of the infoset historically identified by
+    `label` -- not necessarily the representative one, so profile indexing is
+    verified to work from any member, not just the canonical one."""
+    history = _infoset_history(game, label)
+    member = next(iter(game.get_members(gbt.H.path(*history))))
+    return gbt.H.path(*member)
 
 
 @pytest.mark.parametrize(
@@ -163,10 +179,10 @@ def test_profile_indexing_by_player_infoset_action_reference(
     rational_flag: bool,
 ):
     profile = game.mixed_behavior_profile(rational=rational_flag)
-    infoset = games.find_infoset(game, player_label, infoset_label)
-    node = next(iter(infoset.members))
+    history = _infoset_history(game, infoset_label)
+    assert game.get_player(gbt.H.path(*history)) == player_label
     prob = gbt.Rational(prob) if rational_flag else prob
-    assert profile[games.selector_for_node(node)][action_label] == prob
+    assert profile[_member_selector(game, infoset_label)][action_label] == prob
 
 
 @pytest.mark.parametrize(
@@ -228,10 +244,11 @@ def test_profile_indexing_by_selector_reference(
     """profile[selector] and profile[player_label][selector] resolve to the same
     MixedAction."""
     profile = game.mixed_behavior_profile(rational=rational_flag)
-    infoset = games.find_infoset(game, player_label, infoset_label)
-    selector = games.selector_for_node(next(iter(infoset.members)))
+    history = _infoset_history(game, infoset_label)
+    assert game.get_player(gbt.H.path(*history)) == player_label
+    selector = _member_selector(game, infoset_label)
     probs = [gbt.Rational(prob) for prob in probs] if rational_flag else probs
-    expected = dict(zip(infoset.actions, probs, strict=True))
+    expected = dict(zip(game.get_actions(gbt.H.path(*history)), probs, strict=True))
     assert profile[player_label][selector] == expected
     assert profile[selector] == expected
 
@@ -251,7 +268,7 @@ def test_behavior_indexing_rejects_selector_from_different_player(
     """
     profile = game.mixed_behavior_profile()
     other_infoset = games.player_infosets(game, other_player_label)[0]
-    other_selector = games.selector_for_node(next(iter(other_infoset.members)))
+    other_selector = gbt.H.path(*other_infoset)
     with pytest.raises(gbt.MismatchError):
         profile[player_label][other_selector]
 
@@ -278,7 +295,7 @@ def test_profile_indexing_by_player_label_reference(
     if rational_flag:
         behav_data = [[gbt.Rational(prob) for prob in probs] for probs in behav_data]
     expected = [
-        dict(zip(infoset.actions, probs, strict=True))
+        dict(zip(game.get_actions(gbt.H.path(*infoset)), probs, strict=True))
         for infoset, probs in zip(
             games.player_infosets(game, player_label), behav_data, strict=True
         )
@@ -321,8 +338,7 @@ def test_set_probabilities_action(
     """A sparse one-action distribution leaves the infoset's other actions at weight zero."""
     profile = game.mixed_behavior_profile(rational=rational_flag)
     prob = gbt.Rational(prob) if rational_flag else prob
-    node = next(iter(games.find_infoset_in_game(game, infoset_label).members))
-    selector = games.selector_for_node(node)
+    selector = _member_selector(game, infoset_label)
     profile[selector] = {action_label: prob}
     assert profile[selector][action_label] == prob
 
@@ -398,10 +414,10 @@ def test_set_probabilities_infoset(
     profile = game.mixed_behavior_profile(rational=rational_flag)
     if rational_flag:
         probs = [gbt.Rational(p) for p in probs]
-    infoset = games.find_infoset(game, player_label, infoset_label)
-    node = next(iter(infoset.members))
-    selector = games.selector_for_node(node)
-    expected = dict(zip(infoset.actions, probs, strict=True))
+    history = _infoset_history(game, infoset_label)
+    assert game.get_player(gbt.H.path(*history)) == player_label
+    selector = _member_selector(game, infoset_label)
+    expected = dict(zip(game.get_actions(gbt.H.path(*history)), probs, strict=True))
     profile[selector] = expected
     assert profile[selector] == expected
 
@@ -432,7 +448,7 @@ def test_set_probabilities_player_by_label(
     if rational_flag:
         behav_data = [[gbt.Rational(prob) for prob in probs] for probs in behav_data]
     expected = [
-        dict(zip(infoset.actions, probs, strict=True))
+        dict(zip(game.get_actions(gbt.H.path(*infoset)), probs, strict=True))
         for infoset, probs in zip(
             games.player_infosets(game, player_label), behav_data, strict=True
         )
@@ -440,25 +456,26 @@ def test_set_probabilities_player_by_label(
     for infoset, distribution in zip(
         games.player_infosets(game, player_label), expected, strict=True
     ):
-        profile[games.selector_for_node(next(iter(infoset.members)))] = distribution
+        profile[gbt.H.path(*infoset)] = distribution
     assert profile[player_label] == expected
 
 
-def _p1_node(game: gbt.Game):
-    return next(iter(games.player_infosets(game, "Player 1")[0].members))
+def _p1_selector(game: gbt.Game) -> gbt.Selector:
+    """A Selector for Player 1's first information set."""
+    return gbt.H.path(*games.player_infosets(game, "Player 1")[0])
 
 
 def test_behavior_setitem_allows_sparse_distribution():
     game = games.read_from_file("mixed_behavior_game.efg")
     profile = game.mixed_behavior_profile()
-    selector = games.selector_for_node(_p1_node(game))
+    selector = _p1_selector(game)
     profile[selector] = {"U1": 1}
     assert profile[selector] == {"U1": 1, "D1": 0}
 
 
 def test_set_mixed_action_sparse_matches_setitem():
     game = games.read_from_file("mixed_behavior_game.efg")
-    selector = games.selector_for_node(_p1_node(game))
+    selector = _p1_selector(game)
     sparse_profile = game.mixed_behavior_profile()
     sparse_profile.set_mixed_action(selector, {"U1": 1}, sparse=True)
     setitem_profile = game.mixed_behavior_profile()
@@ -469,7 +486,7 @@ def test_set_mixed_action_sparse_matches_setitem():
 def test_set_mixed_action_defaults_to_requiring_every_label():
     game = games.read_from_file("mixed_behavior_game.efg")
     profile = game.mixed_behavior_profile()
-    selector = games.selector_for_node(_p1_node(game))
+    selector = _p1_selector(game)
     with pytest.raises(ValueError, match="exactly one weight"):
         profile.set_mixed_action(selector, {"U1": 1})
 
@@ -478,7 +495,7 @@ def test_set_mixed_action_defaults_to_requiring_every_label():
 def test_setitem_and_set_mixed_action_reject_unknown_action_label(sparse: bool):
     game = games.read_from_file("mixed_behavior_game.efg")
     profile = game.mixed_behavior_profile()
-    selector = games.selector_for_node(_p1_node(game))
+    selector = _p1_selector(game)
     with pytest.raises(ValueError, match="not an action label"):
         profile.set_mixed_action(selector, {"not-an-action": 1}, sparse=sparse)
     with pytest.raises(ValueError, match="not an action label"):
@@ -488,7 +505,7 @@ def test_setitem_and_set_mixed_action_reject_unknown_action_label(sparse: bool):
 def test_behavior_setitem_empty_distribution_is_all_zero_error():
     game = games.read_from_file("mixed_behavior_game.efg")
     profile = game.mixed_behavior_profile()
-    selector = games.selector_for_node(_p1_node(game))
+    selector = _p1_selector(game)
     with pytest.raises(ValueError, match="zero"):
         profile[selector] = {}
 
@@ -497,7 +514,7 @@ def test_behavior_setitem_empty_distribution_is_all_zero_error():
 def test_setitem_and_set_mixed_action_reject_non_mapping(sparse: bool):
     game = games.read_from_file("mixed_behavior_game.efg")
     profile = game.mixed_behavior_profile()
-    selector = games.selector_for_node(_p1_node(game))
+    selector = _p1_selector(game)
     with pytest.raises(TypeError, match="Mapping"):
         profile.set_mixed_action(selector, [1, 0], sparse=sparse)
     with pytest.raises(TypeError, match="Mapping"):
@@ -508,7 +525,7 @@ def test_setitem_and_set_mixed_action_reject_non_mapping(sparse: bool):
 def test_setitem_and_set_mixed_action_reject_uncoercible_weight(sparse: bool):
     game = games.read_from_file("mixed_behavior_game.efg")
     profile = game.mixed_behavior_profile()
-    selector = games.selector_for_node(_p1_node(game))
+    selector = _p1_selector(game)
     full_distribution = {"U1": "abc", "D1": 0}
     with pytest.raises(ValueError, match="convert"):
         profile.set_mixed_action(selector, full_distribution, sparse=sparse)
@@ -520,32 +537,25 @@ def test_behavior_setitem_sparse_rejects_negative_weight():
     """Negativity is checked even for weights given under a sparse distribution."""
     game = games.read_from_file("mixed_behavior_game.efg")
     profile = game.mixed_behavior_profile()
-    selector = games.selector_for_node(_p1_node(game))
+    selector = _p1_selector(game)
     with pytest.raises(ValueError, match="negative"):
         profile[selector] = {"U1": -1}
 
 
 @pytest.mark.parametrize("sparse", [False, True])
-def test_behavior_indexing_rejects_node_and_infoset(sparse: bool):
-    """MixedBehaviorProfile's indexing is Selector-only; a bare Node or Infoset object
-    is rejected, following the same pattern as `Game.get_minimal_subgame`.
+def test_behavior_indexing_rejects_history(sparse: bool):
+    """MixedBehaviorProfile's indexing is Selector-only; a bare `History` tuple is
+    not accepted, following the same pattern as `Game.get_minimal_subgame`.
     """
     game = games.read_from_file("mixed_behavior_game.efg")
     profile = game.mixed_behavior_profile()
-    infoset = games.player_infosets(game, "Player 1")[0]
-    node = next(iter(infoset.members))
+    history = games.player_infosets(game, "Player 1")[0]
     with pytest.raises(TypeError):
-        profile[infoset]
+        profile[history]
     with pytest.raises(TypeError):
-        profile[infoset] = {"U1": 1}
+        profile[history] = {"U1": 1}
     with pytest.raises(TypeError):
-        profile.set_mixed_action(infoset, {"U1": 1}, sparse=sparse)
-    with pytest.raises(TypeError):
-        profile[node]
-    with pytest.raises(TypeError):
-        profile[node] = {"U1": 1}
-    with pytest.raises(TypeError):
-        profile.set_mixed_action(node, {"U1": 1}, sparse=sparse)
+        profile.set_mixed_action(history, {"U1": 1}, sparse=sparse)
 
 
 @pytest.mark.parametrize("rational_flag", [False, True])
@@ -555,7 +565,7 @@ def test_mixed_action_and_behavior_are_frozen_snapshots(rational_flag: bool):
     """
     game = games.read_from_file("mixed_behavior_game.efg")
     profile = game.mixed_behavior_profile(rational=rational_flag)
-    selector = games.selector_for_node(_p1_node(game))
+    selector = _p1_selector(game)
     action_before = profile[selector]
     behavior_before = profile["Player 1"]
     profile[selector] = {"U1": 1, "D1": 0}
@@ -568,7 +578,7 @@ def test_mixed_action_and_behavior_are_frozen_snapshots(rational_flag: bool):
 def test_behavior_copy_mutating_copy_does_not_affect_original(rational_flag: bool):
     game = games.read_from_file("mixed_behavior_game.efg")
     original = game.mixed_behavior_profile(rational=rational_flag)
-    selector = games.selector_for_node(_p1_node(game))
+    selector = _p1_selector(game)
     original_before = dict(original[selector])
     copy = original.copy()
     copy[selector] = {"U1": 1, "D1": 0}
@@ -580,7 +590,7 @@ def test_behavior_copy_mutating_copy_does_not_affect_original(rational_flag: boo
 def test_behavior_copy_mutating_original_does_not_affect_copy(rational_flag: bool):
     game = games.read_from_file("mixed_behavior_game.efg")
     original = game.mixed_behavior_profile(rational=rational_flag)
-    selector = games.selector_for_node(_p1_node(game))
+    selector = _p1_selector(game)
     copy = original.copy()
     copy_before = dict(copy[selector])
     original[selector] = {"U1": 1, "D1": 0}
@@ -598,7 +608,7 @@ def test_as_float_returns_double(rational_flag: bool):
 def test_as_float_converts_rational_probabilities():
     game = games.read_from_file("mixed_behavior_game.efg")
     profile = game.mixed_behavior_profile(rational=True)
-    selector = games.selector_for_node(_p1_node(game))
+    selector = _p1_selector(game)
     profile[selector] = {"U1": "1/3", "D1": "2/3"}
     result = profile.as_float()
     assert dict(result[selector]) == {"U1": pytest.approx(1 / 3), "D1": pytest.approx(2 / 3)}
@@ -607,7 +617,7 @@ def test_as_float_converts_rational_probabilities():
 def test_as_float_is_independent_copy():
     game = games.read_from_file("mixed_behavior_game.efg")
     profile = game.mixed_behavior_profile(rational=False)
-    selector = games.selector_for_node(_p1_node(game))
+    selector = _p1_selector(game)
     result = profile.as_float()
     assert result == profile
     result[selector] = {"U1": 1.0, "D1": 0.0}
@@ -674,14 +684,10 @@ def test_as_float_is_independent_copy():
 def test_realiz_prob_nodes_reference(
     game: gbt.Game, path: list[str], realiz_prob: str | float, rational_flag: bool
 ):
-    # nodes have no labels, so each node is reached by walking the action-label
-    # path from the root (an empty path is the root itself)
+    # a node's History is the tuple of action labels from the root (empty for the root)
     profile = game.mixed_behavior_profile(rational=rational_flag)
     realiz_prob = gbt.Rational(realiz_prob) if rational_flag else realiz_prob
-    node = game.root
-    for action_label in path:
-        node = node.children[action_label]
-    assert profile.realiz_probs[node] == realiz_prob
+    assert profile.realiz_probs[tuple(path)] == realiz_prob
 
 
 @pytest.mark.parametrize(
@@ -697,7 +703,7 @@ def test_infoset_probs_reference(game: gbt.Game, rational_flag: bool, infoset_pr
     profile = game.mixed_behavior_profile(rational=rational_flag)
     for prob, infoset in zip(infoset_probs, games.all_infosets(game), strict=True):
         prob = gbt.Rational(prob) if rational_flag else prob
-        assert profile.infoset_probs[next(iter(infoset.members))] == prob
+        assert profile.infoset_probs[gbt.H.path(*infoset)] == prob
 
 
 @pytest.mark.parametrize(
@@ -735,8 +741,7 @@ def test_absent_minded_infoset_prob(
     game: gbt.Game, infoset_label: str, prob: str | float, rational_flag: bool
 ):
     profile = game.mixed_behavior_profile(rational=rational_flag)
-    node = next(iter(games.find_infoset_in_game(game, infoset_label).members))
-    ip = profile.infoset_probs[node]
+    ip = profile.infoset_probs[_member_selector(game, infoset_label)]
     assert ip == (gbt.Rational(prob) if rational_flag else prob)
 
 
@@ -746,8 +751,8 @@ def test_nature_rooted_game_root_reached_with_certainty(rational_flag: bool):
     game = gbt.catalog.load("journals/geb/gilboa1997/fig2")
     profile = game.mixed_behavior_profile(rational=rational_flag)
     one = gbt.Rational(1) if rational_flag else 1.0
-    assert profile.realiz_probs[game.root] == one
-    assert profile.infoset_probs[game.root] == one
+    assert profile.realiz_probs[()] == one
+    assert profile.infoset_probs[gbt.H.path()] == one
 
 
 @pytest.mark.parametrize(
@@ -763,7 +768,7 @@ def test_infoset_values_reference(game: gbt.Game, rational_flag: bool, infoset_v
     profile = game.mixed_behavior_profile(rational=rational_flag)
     for payoff, infoset in zip(infoset_values, games.all_infosets(game), strict=True):
         payoff = gbt.Rational(payoff) if rational_flag else payoff
-        assert profile.infoset_values[next(iter(infoset.members))] == payoff
+        assert profile.infoset_values[gbt.H.path(*infoset)] == payoff
 
 
 @pytest.mark.parametrize(
@@ -786,8 +791,11 @@ def test_infoset_values_reference(game: gbt.Game, rational_flag: bool, infoset_v
 def test_action_values_reference(game: gbt.Game, rational_flag: bool, action_values: tuple):
     profile = game.mixed_behavior_profile(rational=rational_flag)
     for values_for_infoset, infoset in zip(action_values, games.all_infosets(game), strict=True):
-        infoset_action_values = profile.action_values[next(iter(infoset.members))]
-        for value, action in zip(values_for_infoset, infoset.actions, strict=True):
+        selector = gbt.H.path(*infoset)
+        infoset_action_values = profile.action_values[selector]
+        for value, action in zip(
+            values_for_infoset, game.get_actions(selector), strict=True
+        ):
             value = gbt.Rational(value) if rational_flag else value
             assert infoset_action_values[action] == value
 
@@ -807,11 +815,12 @@ def test_action_regret_consistency(game: gbt.Game, rational_flag: bool):
     profile = game.mixed_behavior_profile(rational=rational_flag)
     for player in game.players:
         for infoset in games.player_infosets(game, player):
-            node = next(iter(infoset.members))
-            for action in infoset.actions:
-                assert profile.action_regrets[node][action] == max(
-                    profile.action_values[node][a] for a in infoset.actions
-                ) - profile.action_values[node][action]
+            selector = gbt.H.path(*infoset)
+            actions = game.get_actions(selector)
+            for action in actions:
+                assert profile.action_regrets[selector][action] == max(
+                    profile.action_values[selector][a] for a in actions
+                ) - profile.action_values[selector][action]
 
 
 @pytest.mark.parametrize(
@@ -829,10 +838,10 @@ def test_infoset_regret_consistency(game: gbt.Game, rational_flag: bool):
     profile = game.mixed_behavior_profile(rational=rational_flag)
     for player in game.players:
         for infoset in games.player_infosets(game, player):
-            node = next(iter(infoset.members))
-            assert profile.infoset_regrets[node] == max(
-                profile.action_values[node][a] for a in infoset.actions
-            ) - profile.infoset_values[node]
+            selector = gbt.H.path(*infoset)
+            assert profile.infoset_regrets[selector] == max(
+                profile.action_values[selector][a] for a in game.get_actions(selector)
+            ) - profile.infoset_values[selector]
 
 
 @pytest.mark.parametrize(
@@ -868,7 +877,7 @@ def test_agent_max_regret_consistency(game: gbt.Game, rational_flag: bool):
     profile = game.mixed_behavior_profile(rational=rational_flag)
     infoset_regrets = profile.infoset_regrets
     assert profile.agent_max_regret() == max(
-        infoset_regrets[next(iter(infoset.members))] for infoset in games.all_infosets(game)
+        infoset_regrets[gbt.H.path(*infoset)] for infoset in games.all_infosets(game)
     )
 
 
@@ -913,34 +922,35 @@ def test_vectorized_quantities_consistency(game: gbt.Game, rational_flag: bool):
     for player in game.players:
         player_node_values = node_values[player]
         assert isinstance(player_node_values, gbt.NodeValueVector)
-        assert player_node_values[game.root] == payoffs[player]
+        assert player_node_values[()] == payoffs[player]
 
         for infoset in games.player_infosets(game, player):
-            node = next(iter(infoset.members))
-            infoset_action_values = action_values[node]
-            infoset_action_regrets = action_regrets[node]
+            selector = gbt.H.path(*infoset)
+            actions = game.get_actions(selector)
+            infoset_action_values = action_values[selector]
+            infoset_action_regrets = action_regrets[selector]
             assert isinstance(infoset_action_values, gbt.ActionValueVector)
             assert isinstance(infoset_action_regrets, gbt.ActionRegretVector)
 
-            best_response_value = max(infoset_action_values[a] for a in infoset.actions)
-            assert infoset_regrets[node] == best_response_value - infoset_values[node]
-            for action in infoset.actions:
+            best_response_value = max(infoset_action_values[a] for a in actions)
+            assert infoset_regrets[selector] == best_response_value - infoset_values[selector]
+            for action in actions:
                 assert (
                     infoset_action_regrets[action]
                     == best_response_value - infoset_action_values[action]
                 )
 
-    for node in game.nodes:
-        if node.is_terminal:
+    for history in game.get_histories(gbt.H.after()):
+        if not game.get_actions(gbt.H.path(*history)):
             continue
-        if infoset_probs[node] == 0:
-            assert beliefs[node] is None
+        if infoset_probs[gbt.H.path(*history)] == 0:
+            assert beliefs[history] is None
         else:
-            assert beliefs[node] is not None
+            assert beliefs[history] is not None
 
     # equal to an equivalent plain dict or same-type vector, but never to a vector of a
     # different quantity, even where the underlying numbers happen to coincide
-    expected = {n: realiz_probs[n] for n in game.nodes}
+    expected = dict(realiz_probs)
     assert realiz_probs == expected
     assert realiz_probs == gbt.RealizProbVector(expected)
     assert realiz_probs != beliefs
@@ -1013,8 +1023,11 @@ def test_action_regrets_reference(
     if action_probs:
         _set_action_probs(profile, action_probs, rational_flag)
     for regrets_for_infoset, infoset in zip(action_regrets, games.all_infosets(game), strict=True):
-        infoset_action_regrets = profile.action_regrets[next(iter(infoset.members))]
-        for regret, action in zip(regrets_for_infoset, infoset.actions, strict=True):
+        selector = gbt.H.path(*infoset)
+        infoset_action_regrets = profile.action_regrets[selector]
+        for regret, action in zip(
+            regrets_for_infoset, game.get_actions(selector), strict=True
+        ):
             regret = gbt.Rational(regret) if rational_flag else regret
             assert abs(infoset_action_regrets[action] - regret) <= tol
 
@@ -1036,16 +1049,17 @@ def test_martingale_property_of_node_value(game: gbt.Game, rational_flag: bool):
     profile = game.mixed_behavior_profile(rational=rational_flag)
     realiz_probs = profile.realiz_probs
     node_values = profile.node_values
-    for node in game.nodes:
-        if node.is_terminal or node.player == "Chance":
+    for history in game.get_histories(gbt.H.after()):
+        player = game.get_player(gbt.H.path(*history))
+        if player is None or player == "Chance":
             continue
         expected_val = 0
-        node_prob = realiz_probs[node]
-        player_node_values = node_values[node.player]
-        for child in node.children:
-            prob = realiz_probs[child] / node_prob
-            expected_val += prob * player_node_values[child]
-        assert player_node_values[node] == expected_val
+        node_prob = realiz_probs[history]
+        player_node_values = node_values[player]
+        for child_history in games.children_histories(game, history):
+            prob = realiz_probs[child_history] / node_prob
+            expected_val += prob * player_node_values[child_history]
+        assert player_node_values[history] == expected_val
 
 
 @pytest.mark.parametrize(
@@ -1064,7 +1078,7 @@ def test_node_value_consistency(game: gbt.Game, rational_flag: bool):
     node_values = profile.node_values
     payoffs = profile.payoffs
     for player in game.players:
-        assert node_values[player][game.root] == payoffs[player]
+        assert node_values[player][()] == payoffs[player]
 
 
 @pytest.mark.parametrize(
@@ -1361,15 +1375,11 @@ def test_node_belief_reference(
     value: str | float,
     rational_flag: bool,
 ):
-    # nodes have no labels, so each belief node is reached by walking the
-    # action-label path from the root (an empty path is the root itself)
+    # a node's History is the tuple of action labels from the root (empty for the root)
     profile = game.mixed_behavior_profile(rational=rational_flag)
     _set_action_probs(profile, probs, rational_flag)
-    node = game.root
-    for action_label in path:
-        node = node.children[action_label]
     value = gbt.Rational(value) if rational_flag else value
-    assert abs(profile.beliefs[node] - value) <= tol
+    assert abs(profile.beliefs[tuple(path)] - value) <= tol
 
 
 @pytest.mark.parametrize(
@@ -1385,7 +1395,9 @@ def test_infoset_value_error_with_chance_player_infoset(game: gbt.Game, rational
     """
     chance_node = game.get_events()[0]
     with pytest.raises(KeyError):
-        game.mixed_behavior_profile(rational=rational_flag).infoset_values[chance_node]
+        game.mixed_behavior_profile(rational=rational_flag).infoset_values[
+            gbt.H.path(*chance_node)
+        ]
 
 
 @pytest.mark.parametrize(
@@ -1401,16 +1413,19 @@ def test_action_value_error_with_chance_player_action(game: gbt.Game, rational_f
     """
     chance_node = game.get_events()[0]
     with pytest.raises(KeyError):
-        game.mixed_behavior_profile(rational=rational_flag).action_values[chance_node]
+        game.mixed_behavior_profile(rational=rational_flag).action_values[
+            gbt.H.path(*chance_node)
+        ]
 
 
-def _all_node_actions(game: gbt.Game) -> list[tuple[gbt.Node, str]]:
-    """All (node, action label) pairs across every personal player's information sets."""
+def _all_node_actions(game: gbt.Game) -> list[tuple[gbt.Selector, str]]:
+    """All (selector, action label) pairs across every personal player's information sets."""
     return [
-        (node, action)
+        (selector, action)
         for player in game.players
-        for node in game.get_infosets(player)
-        for action in node.actions
+        for history in game.get_infosets(player)
+        for selector in [gbt.H.path(*history)]
+        for action in game.get_actions(selector)
     ]
 
 
@@ -1475,7 +1490,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_2A_doub,
             False,
             lambda x, y: x.beliefs[y],
-            lambda x: x.nodes,
+            lambda x: x.get_histories(gbt.H.after()),
         ),
         (
             games.read_from_file("mixed_behavior_game.efg"),
@@ -1483,7 +1498,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_2A_rat,
             True,
             lambda x, y: x.beliefs[y],
-            lambda x: x.nodes,
+            lambda x: x.get_histories(gbt.H.after()),
         ),
         (
             games.create_stripped_down_poker_efg(),
@@ -1491,7 +1506,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_2B_doub,
             False,
             lambda x, y: x.beliefs[y],
-            lambda x: x.nodes,
+            lambda x: x.get_histories(gbt.H.after()),
         ),
         (
             games.create_stripped_down_poker_efg(),
@@ -1499,7 +1514,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_2A_rat,
             True,
             lambda x, y: x.beliefs[y],
-            lambda x: x.nodes,
+            lambda x: x.get_histories(gbt.H.after()),
         ),
         ######################################################################################
         # realiz_prob (at nodes)
@@ -1509,7 +1524,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_2A_doub,
             False,
             lambda x, y: x.realiz_probs[y],
-            lambda x: x.nodes,
+            lambda x: x.get_histories(gbt.H.after()),
         ),
         (
             games.read_from_file("mixed_behavior_game.efg"),
@@ -1517,7 +1532,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_2A_rat,
             True,
             lambda x, y: x.realiz_probs[y],
-            lambda x: x.nodes,
+            lambda x: x.get_histories(gbt.H.after()),
         ),
         (
             games.create_stripped_down_poker_efg(),
@@ -1525,7 +1540,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_2B_doub,
             False,
             lambda x, y: x.realiz_probs[y],
-            lambda x: x.nodes,
+            lambda x: x.get_histories(gbt.H.after()),
         ),
         (
             games.create_stripped_down_poker_efg(),
@@ -1533,7 +1548,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_2A_rat,
             True,
             lambda x, y: x.realiz_probs[y],
-            lambda x: x.nodes,
+            lambda x: x.get_histories(gbt.H.after()),
         ),
         ######################################################################################
         # infoset_prob
@@ -1542,7 +1557,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_1A_doub,
             PROBS_2A_doub,
             False,
-            lambda x, y: x.infoset_probs[next(iter(y.members))],
+            lambda x, y: x.infoset_probs[gbt.H.path(*y)],
             lambda x: games.all_infosets(x),
         ),
         (
@@ -1550,7 +1565,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_1A_rat,
             PROBS_2A_rat,
             True,
-            lambda x, y: x.infoset_probs[next(iter(y.members))],
+            lambda x, y: x.infoset_probs[gbt.H.path(*y)],
             lambda x: games.all_infosets(x),
         ),
         (
@@ -1558,7 +1573,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_1B_doub,
             PROBS_2B_doub,
             False,
-            lambda x, y: x.infoset_probs[next(iter(y.members))],
+            lambda x, y: x.infoset_probs[gbt.H.path(*y)],
             lambda x: games.all_infosets(x),
         ),
         (
@@ -1566,7 +1581,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_1A_rat,
             PROBS_2A_rat,
             True,
-            lambda x, y: x.infoset_probs[next(iter(y.members))],
+            lambda x, y: x.infoset_probs[gbt.H.path(*y)],
             lambda x: games.all_infosets(x),
         ),
         ######################################################################################
@@ -1576,7 +1591,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_1A_doub,
             PROBS_2A_doub,
             False,
-            lambda x, y: x.infoset_values[next(iter(y.members))],
+            lambda x, y: x.infoset_values[gbt.H.path(*y)],
             lambda x: games.all_infosets(x),
         ),
         (
@@ -1584,7 +1599,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_1A_rat,
             PROBS_2A_rat,
             True,
-            lambda x, y: x.infoset_values[next(iter(y.members))],
+            lambda x, y: x.infoset_values[gbt.H.path(*y)],
             lambda x: games.all_infosets(x),
         ),
         (
@@ -1592,7 +1607,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_1B_doub,
             PROBS_2B_doub,
             False,
-            lambda x, y: x.infoset_values[next(iter(y.members))],
+            lambda x, y: x.infoset_values[gbt.H.path(*y)],
             lambda x: games.all_infosets(x),
         ),
         (
@@ -1600,7 +1615,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_1A_rat,
             PROBS_2A_rat,
             True,
-            lambda x, y: x.infoset_values[next(iter(y.members))],
+            lambda x, y: x.infoset_values[gbt.H.path(*y)],
             lambda x: games.all_infosets(x),
         ),
         ######################################################################################
@@ -1679,7 +1694,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_2A_doub,
             False,
             lambda x, y: x.node_values[y[0]][y[1]],
-            lambda x: list(product(x.players, x.nodes)),
+            lambda x: list(product(x.players, x.get_histories(gbt.H.after()))),
         ),
         (
             games.read_from_file("mixed_behavior_game.efg"),
@@ -1687,7 +1702,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_2A_rat,
             True,
             lambda x, y: x.node_values[y[0]][y[1]],
-            lambda x: list(product(x.players, x.nodes)),
+            lambda x: list(product(x.players, x.get_histories(gbt.H.after()))),
         ),
         (
             games.create_stripped_down_poker_efg(),
@@ -1695,7 +1710,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_2B_doub,
             False,
             lambda x, y: x.node_values[y[0]][y[1]],
-            lambda x: list(product(x.players, x.nodes)),
+            lambda x: list(product(x.players, x.get_histories(gbt.H.after()))),
         ),
         (
             games.create_stripped_down_poker_efg(),
@@ -1703,7 +1718,7 @@ PROBS_2B_doub = (1.0, 0.0, 1.0, 0.0, 1.0, 0.0)
             PROBS_2A_rat,
             True,
             lambda x, y: x.node_values[y[0]][y[1]],
-            lambda x: list(product(x.players, x.nodes)),
+            lambda x: list(product(x.players, x.get_histories(gbt.H.after()))),
         ),
         ######################################################################################
         # agent_liap_value (of profile, hence [1] for objects_to_test,
@@ -1894,8 +1909,8 @@ def test_specific_profile(game: gbt.Game, rational_flag: bool, data: list):
     profile = game.mixed_behavior_profile(rational=rational_flag, data=data)
     flattened = iter([k for i in data for j in i for k in j])
     for infoset in games.all_infosets(game):
-        selector = games.selector_for_node(next(iter(infoset.members)))
-        for action in infoset.actions:
+        selector = gbt.H.path(*infoset)
+        for action in game.get_actions(selector):
             prob = next(flattened)
             assert profile[selector][action] == (gbt.Rational(prob) if rational_flag else prob)
 
@@ -1983,29 +1998,28 @@ def test_undefined_action_value():
     """Test that undefined action values return `None`."""
     game = gbt.catalog.load("journals/ijgt/selten1975/fig1")
     *_, p3 = game.players
-    infoset = games.player_infosets(game, p3)[0]
-    node = next(iter(infoset.members))
-    action = next(iter(infoset.actions))
+    selector = gbt.H.path(*games.player_infosets(game, p3)[0])
+    action = game.get_actions(selector)[0]
     for rat in [False, True]:
         profile = game.mixed_behavior_profile([[[1, 0]], [[1, 0]], [[1, 0]]], rational=rat)
-        assert profile.action_values[node][action] is None
+        assert profile.action_values[selector][action] is None
 
 
 def test_undefined_belief():
     """Test that undefined beliefs return `None`."""
     game = gbt.catalog.load("journals/ijgt/selten1975/fig1")
     *_, p3 = game.players
-    node = next(iter(games.player_infosets(game, p3)[0].members))
+    history = games.player_infosets(game, p3)[0]
     for rat in [False, True]:
         profile = game.mixed_behavior_profile([[[1, 0]], [[1, 0]], [[1, 0]]], rational=rat)
-        assert profile.beliefs[node] is None
+        assert profile.beliefs[history] is None
 
 
 def test_undefined_infoset_value():
     """Test that undefined infoset values return `None`."""
     game = gbt.catalog.load("journals/ijgt/selten1975/fig1")
     *_, p3 = game.players
-    node = next(iter(games.player_infosets(game, p3)[0].members))
+    selector = gbt.H.path(*games.player_infosets(game, p3)[0])
     for rat in [False, True]:
         profile = game.mixed_behavior_profile([[[1, 0]], [[1, 0]], [[1, 0]]], rational=rat)
-        assert profile.infoset_values[node] is None
+        assert profile.infoset_values[selector] is None
