+++ tests/test_game_resolve.py
@@ -9,14 +9,11 @@ from . import games
 
 
 def _test_valid_resolutions(collection: list, resolver: typing.Callable) -> None:
-    """Generic function to exercise resolving objects as themselves or via existing labels."""
+    """Generic function to exercise resolving objects via their existing labels."""
     for label, objects in itertools.groupby(
-            sorted(collection, key=lambda x: x.label), lambda x: x.label
+            sorted(collection, key=lambda x: x._label), lambda x: x._label
     ):
         objects = list(objects)
-        # Objects resolve to themselves
-        for obj in objects:
-            assert obj == resolver(obj, "test")
         # Ambiguous labels raise ValueError
         if len(objects) > 1:
             with pytest.raises(ValueError):
@@ -32,7 +29,7 @@ def _test_valid_resolutions(collection: list, resolver: typing.Callable) -> None
     ]
 )
 def test_resolve_node(game: gbt.Game) -> None:
-    _test_valid_resolutions(game.nodes,
+    _test_valid_resolutions(games.all_nodes(game),
                             lambda label, fn: game._resolve_node(label, fn))
 
 
@@ -49,12 +46,12 @@ def test_resolve_node_invalid(game: gbt.Game, node: str, exception: BaseExceptio
         game._resolve_node(node, "test_resolve_node_invalid")
 
 
-def test_resolve_node_mismatch():
-    """A `Node` from a different game raises `MismatchError`."""
-    game1 = gbt.Game.new_tree()
-    game2 = games.read_from_file("sample_extensive_game.efg")
-    with pytest.raises(gbt.MismatchError):
-        game1._resolve_node(game2.root, "test_resolve_node_mismatch")
+def test_resolve_node_rejects_node():
+    """A bare `Node` -- even one belonging to this same game -- is not accepted;
+    `_resolve_node` only resolves a `Selector`, `History` tuple, or label `str`."""
+    game = games.read_from_file("sample_extensive_game.efg")
+    with pytest.raises(TypeError):
+        game._resolve_node(games.node_at_history(game, ()), "test_resolve_node_rejects_node")
 
 
 @pytest.mark.parametrize(
@@ -64,16 +61,18 @@ def test_resolve_node_mismatch():
     ]
 )
 def test_resolve_infoset(game: gbt.Game) -> None:
-    """`_resolve_infoset` resolves a Node to itself, or a node's label to the node
-    with that label, validating that it currently belongs to a personal player's
-    information set; this holds for every member node, not just the representative
-    `Game.get_infosets` returns."""
+    """`_resolve_infoset` resolves a Selector to the node it identifies, or a
+    node's label to that node, validating that it currently belongs to a
+    personal player's information set; this holds for every member node, not
+    just the representative `Game.get_infosets` returns."""
     for player in game.players:
-        for node in game.get_infosets(player):
-            for member in node.members:
-                assert game._resolve_infoset(member, "test") == member
-                if member.label:
-                    assert game._resolve_infoset(member.label, "test") == member
+        for history in game.get_infosets(player):
+            for member in game.get_members(gbt.H.path(*history)):
+                resolved = game._resolve_infoset(gbt.H.path(*member), "test")
+                assert games._node_history(resolved) == member
+                if resolved._label:
+                    resolved_by_label = game._resolve_infoset(resolved._label, "test")
+                    assert games._node_history(resolved_by_label) == member
 
 
 @pytest.mark.parametrize(
