+++ tests/test_players.py
@@ -305,7 +305,7 @@ def test_player_sequence_count():
     game = gbt.catalog.load("books/myerson1991/fig2_1")
     for player in game.players:
         action_count = sum(
-            len(node.actions) for node in game.get_infosets(player)
+            len(game.get_actions(gbt.H.path(*history))) for history in game.get_infosets(player)
         )
         assert len(game.get_sequences(player)) == action_count + 1
 
@@ -317,8 +317,8 @@ def test_player_sequence_actions():
     reference = (
         set(
             (action, )
-            for node in game.get_infosets(player)
-            for action in node.actions
+            for history in game.get_infosets(player)
+            for action in game.get_actions(gbt.H.path(*history))
         ) |
         {tuple()}
     )
